include: soc: camrtc: address cert int30-c defects
Jira CAMASIL-1298

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2318845
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit

+++ include/soc/tegra/camrtc-isp5-tiling.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017-2019, NVIDIA CORPORATION.  All rights reserved.
+ * Copyright (c) 2017-2020, NVIDIA CORPORATION.  All rights reserved.
  *
  * NVIDIA CORPORATION and its licensors retain all intellectual property
  * and proprietary rights in and to this software, related documentation
@@ -55,7 +55,9 @@ static inline uint16_t isp5_align_up(const uint16_t val, const uint16_t alignmen
 
 static inline uint16_t isp5_div_round_up(const uint16_t x, const uint16_t y)
 {
-	return (x + (y - 1U)) / y;
+	//coverity[cert_int30_c_violation] # x & y are uint16_t
+	return (uint16_t)(((uint32_t)x + ((uint32_t)y - 1U)) / (uint32_t)y);
+
 }
 
 /**
@@ -102,6 +104,7 @@ static bool isp5_find_tile_width(const struct isp5_program * const prg,
 		const uint16_t pixels_left = img_width - max_width_first - max_width_last;
 		const uint16_t middle_tiles = isp5_div_round_up(pixels_left,
 					      isp5_min_u16(max_width_middle, max_width_first));
+		//coverity[cert_int30_c_violation] # tile_count < image width
 		tile_count += middle_tiles;
 	}
 
@@ -121,6 +124,7 @@ static bool isp5_find_tile_width(const struct isp5_program * const prg,
 					prg->overfetch.pru_ovf_h, alignment) -
 					prg_right) + prg->overfetch.pru_ovf_h);
 	uint16_t middle_width = (tile_count > 2U) ? isp5_min_u16(max_width_middle, tile_width) : 0U;
+	//coverity[cert_int30_c_violation] # tile dimensions <= image dimensions
 	uint16_t last_width = img_width - first_width - ((tile_count - 2U) * middle_width);
 
 	/*
@@ -135,7 +139,9 @@ static bool isp5_find_tile_width(const struct isp5_program * const prg,
 	if (last_width < min_width) {
 		const uint16_t corr = isp5_align_up(min_width-last_width, alignment);
 
+		//coverity[cert_int30_c_violation] # corr <= tile width <= image width
 		first_width -= corr;
+		//coverity[cert_int30_c_violation] # corr <= tile width <= image width
 		last_width  += corr;
 	} else if (last_width > max_width_last) {
 		/* Try first increasing middle tile width */
@@ -145,14 +151,19 @@ static bool isp5_find_tile_width(const struct isp5_program * const prg,
 			const uint16_t middle_corr = isp5_min_u16(max_middle_corr,
 					    isp5_align_up(isp5_div_round_up(corr, tile_count - 2U),
 					    alignment));
+			//coverity[cert_int30_c_violation] # middle_corr <= tile width <= image width
 			middle_width += middle_corr;
+			/* middle_corr <= tile width <= image width, tile_count < image width */
+			//coverity[cert_int30_c_violation] # see above comment
 			last_width -= middle_corr * (tile_count - 2U);
 		}
 
 		if (last_width > max_width_last) {
 			const uint16_t first_corr = isp5_align_up(last_width -
 								  max_width_last, alignment);
+			//coverity[cert_int30_c_violation] # first_corr <= tile width <= image width
 			first_width += first_corr;
+			//coverity[cert_int30_c_violation] # first_corr <= tile width <= image width
 			last_width -= first_corr;
 		}
 	} else {
@@ -227,8 +238,16 @@ static bool isp5_find_tile_width_dpcm(const struct isp5_program * const prg,
 	tiling->tiles_in_slice = 1U + isp5_div_round_up(cd->surface_configs.mr_width -
 						cd->surface_configs.chunk_width_first,
 						cd->surface_configs.chunk_width_middle);
+	/*
+	 * CERT INT30-C deviations:
+	 * Tile properties are governed by image properties.
+	 * tiles_in_slice < image width, tile_width_middle <= image width
+	 * tile_width_first <= image width.
+	 */
+	//coverity[cert_int30_c_violation] # see above comment
 	const uint16_t last_width = cd->surface_configs.mr_width -
 					tiling->tile_width_first -
+					//coverity[cert_int30_c_violation] # see above comment
 					((tiling->tiles_in_slice - 1U) * tiling->tile_width_middle);
 
 	const uint16_t max_width_last = ISP5_MAX_TILE_WIDTH - prg->overfetch.left;
@@ -258,12 +277,14 @@ static bool isp5_find_slice_height(const uint16_t img_height,
 		uint16_t slice_height = ISP5_MAX_SLICE_HEIGHT;
 		const uint16_t slice_count = isp5_div_round_up(img_height, ISP5_MAX_SLICE_HEIGHT);
 		const uint16_t last_height = img_height -
+					//coverity[cert_int30_c_violation] # slice_count >= 1U always
 					(ISP5_MAX_SLICE_HEIGHT * (slice_count - 1U));
 
 		if (last_height < ISP5_MIN_SLICE_HEIGHT) {
 			const uint16_t corr = ISP5_MIN_SLICE_HEIGHT - last_height;
 			const uint16_t slice_corr =
 				       isp5_align_up(isp5_div_round_up(corr, slice_count - 1U), 2U);
+			//coverity[cert_int30_c_violation] # slice_corr < slice_height always
 			slice_height -= slice_corr;
 		}
